Convert Float to Double

+++ ambe.c
@@ -117,13 +117,13 @@ mbe_dequantizeAmbeParms (mbe_parms * cur_mp, mbe_parms * prev_mp, const int *b, 
         f0 = AmbeW0table[b0];
       cur_mp->w0 = f0 * 2.0 * M_PI;
       // w0 from patent filings
-      //f0 = powf (2, ((float) b0 + (float) 195.626) / -(float) 45.368);
-      //cur_mp->w0 = f0 * (float) 2 *M_PI;
+      //f0 = powf (2, ((double) b0 + (double) 195.626) / -(double) 45.368);
+      //cur_mp->w0 = f0 * (double) 2 *M_PI;
     }
 
   unvc = 0.2046 / sqrt (cur_mp->w0);
-  //unvc = (float) 1;
-  //unvc = (float) 0.2046 / sqrtf (f0);
+  //unvc = (double) 1;
+  //unvc = (double) 0.2046 / sqrtf (f0);
 
   // decode L
   if (silence == 0)
@@ -135,7 +135,7 @@ mbe_dequantizeAmbeParms (mbe_parms * cur_mp, mbe_parms * prev_mp, const int *b, 
       else
         L = (int)AmbeLtable[b0];
       // L formula form patent filings
-      //L=(int)((float)0.4627 / f0);
+      //L=(int)((double)0.4627 / f0);
       cur_mp->L = L;
     }
 
@@ -145,7 +145,7 @@ mbe_dequantizeAmbeParms (mbe_parms * cur_mp, mbe_parms * prev_mp, const int *b, 
       // jl from specification document
       jl = (int) ( l * 16.0 * f0);
       // jl from patent filings?
-      //jl = (int)(((float)l * (float)16.0 * f0) + 0.25);
+      //jl = (int)(((double)l * (double)16.0 * f0) + 0.25);
 
       if (silence == 0)
         {
@@ -407,14 +407,14 @@ mbe_dequantizeAmbeParms (mbe_parms * cur_mp, mbe_parms * prev_mp, const int *b, 
       Sum42 += Tl[l];
     }
   Sum42 = Sum42 / cur_mp->L;
-  BigGamma = cur_mp->gamma - ( 0.5f * (log (cur_mp->L) / logf (2.0f))) - Sum42;
-  //BigGamma=cur_mp->gamma - ((float)0.5 * log((float)cur_mp->L)) - Sum42;
+  BigGamma = cur_mp->gamma - ( 0.5 * (log ((double)cur_mp->L) / log (2.0))) - Sum42;
+  //BigGamma=cur_mp->gamma - (0.5 * log((double)cur_mp->L)) - Sum42;
 
   // Part 3
   for (l = 1; l <= cur_mp->L; l++)
     {
-      c1 = ((float) 0.65 * ((float) 1 - deltal[l]) * prev_mp->log2Ml[intkl[l]]);
-      c2 = ((float) 0.65 * deltal[l] * prev_mp->log2Ml[intkl[l] + 1]);
+      c1 = ( 0.65 * ( 1.0 - deltal[l]) * prev_mp->log2Ml[intkl[l]]);
+      c2 = ( 0.65 * deltal[l] * prev_mp->log2Ml[intkl[l] + 1]);
       cur_mp->log2Ml[l] = Tl[l] + c1 + c2 - Sum43 + BigGamma;
       // inverse log to generate spectral amplitudes
       if (cur_mp->Vl[l] == 1)
